✨(frontend) add TranslatableForm to manage translatable forms
Currently, translatable content is no longer suitable. We transform it into a
TranslatableForm by adding specific logic to language changes.
